encoders/ffmpeg: Rename files to match naming scheme
