+++ source/encoders/encoder-ffmpeg.cpp
@@ -19,7 +19,7 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 // SOFTWARE.
 
-#include "ffmpeg-encoder.hpp"
+#include "encoder-ffmpeg.hpp"
 #include "strings.hpp"
 #include <sstream>
 #include "codecs/hevc.hpp"
